Reorganize Application layer: rename SelectPatchResolver, group Subject exceptions
- Rename SelectPatchResolver → SelectStatementResolver. The "Patch" name
  was tied to the old PATCH endpoint; the resolver itself just walks an
  incoming statements array and resolves select-typed values to canonical
  option IDs, regardless of which write endpoint called it.
- Move Subject exceptions to src/Application/Subject/Exception/, mirroring
  the src/Application/Query/Exception/ layout introduced by the public
  query API. Keeps domain-grouped exceptions together as both areas grow.

No behavior change.
